requirements.txt
asgiref==3.4.1
certifi==2021.5.30
cycler==0.10.0
dj-database-url==0.5.0
dj-static==0.0.6
Django==3.2
django-debug-toolbar==3.2.2
django-querysetsequence==0.14
kiwisolver==1.3.2
matplotlib==3.4.3
numpy==1.21.2
Pillow==8.3.1
pyparsing==2.4.7
python-dateutil==2.8.2
python-decouple==3.4
pytz==2021.1
six==1.16.0
sqlparse==0.4.1
static3==0.7.0
